+++ composer.json
@@ -21,6 +21,7 @@
         "firebase/php-jwt": "^5.0",
         "gumlet/php-image-resize": "^1.9",
         "symfony/var-dumper": "^4.2",
+        "symfony/var-exporter": "^5.2",
         "symfony/console": "^4.3",
         "vlucas/phpdotenv": "^2.4",
         "dflydev/dot-access-data": "^2.0",
